Move stuff into utils file in create-emotion-styled
